backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = ""
authors = [
    {name = "null",email = "you@example.com"}
]
packages = [
    { include = "app", from = "." }
]
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.116.1,<0.117.0",
    "uvicorn[standard]>=0.35.0,<0.36.0",
    "sqlalchemy>=2.0.43,<3.0.0",
    "pymysql>=1.1.2,<2.0.0",
    "pymongo>=4.14.1,<5.0.0",
    "neo4j>=5.28.2,<6.0.0",
    "python-dotenv>=1.1.1,<2.0.0",
    "alembic>=1.16.5,<2.0.0",
    "email-validator>=2.1.1,<3.0.0",
    "python-jose[cryptography]>=3.3.0,<4.0.0",
    "passlib[bcrypt]>=1.7.4,<2.0.0",
    "python-multipart>=0.0.5,<1.0.0",
    "pydantic (>=2.11.7,<3.0.0)",
    "pydantic-settings (>=2.10.1,<3.0.0)"
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
